refactor(hooks): rename typings to types
