+++ packages/hooks/index.d.ts
@@ -1,9 +1,9 @@
-export { default as useEffectAsync } from './typings/useEffectAsync';
-export { default as useMount } from './typings/useMount';
-export { default as useTimeout } from './typings/useTimeout';
-export { default as useToggle } from './typings/useToggle';
-export { default as useCurrentRegion } from './typings/useCurrentRegion';
-export { default as useCurrentUser } from './typings/useCurrentUser';
-export { default as useProviders } from './typings/useProviders';
-export { default as usePermissions } from './typings/usePermissions';
-export { default as useOrganizations } from './typings/useOrganizations';
+export { default as useEffectAsync } from './types/useEffectAsync';
+export { default as useMount } from './types/useMount';
+export { default as useTimeout } from './types/useTimeout';
+export { default as useToggle } from './types/useToggle';
+export { default as useCurrentRegion } from './types/useCurrentRegion';
+export { default as useCurrentUser } from './types/useCurrentUser';
+export { default as useProviders } from './types/useProviders';
+export { default as usePermissions } from './types/usePermissions';
+export { default as useOrganizations } from './types/useOrganizations';
