+++ packages/hooks/package.json
@@ -4,6 +4,7 @@
   "author": "Kyle Gray <kyle.gray@availity.com>",
   "description": "A group of pre-built hooks that are common in most apps",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity"
@@ -26,16 +27,14 @@
   "devDependencies": {
     "@availity/api-axios": "^5.5.14",
     "@availity/api-core": "^7.0.2",
-    "@availity/localstorage-core": "^3.0.0",
     "axios": "^0.21.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "peerDependencies": {
     "@availity/api-axios": "^5.2.1",
-    "@availity/api-core": "^5.2.1",
-    "@availity/localstorage-core": "^2.6.1",
-    "axios": "^0.16.0 || ^0.17.0 || ^0.18.0 || ^0.19.0",
+    "@availity/api-core": "^7.0.2",
+    "axios": "^0.21.1",
     "react": ">=16.8.0"
   }
 }
