+++ packages/hooks/types/usePermissions.d.ts
